Add files via upload
This is a .sql file that I created based on a streaming services dataset available from Kaggle: https://www.kaggle.com/datasets/shivamb/netflix-shows. We will use this data for a class project in which we will decide which streaming service has the most movies and shows that we like.